Codes/PollenTrack- codes KM v0/requirements.txt
picamera2>=0.3.17
RPi.GPIO>=0.7.0
datetime>=5.4
numpy>=1.26.4
opencv-python>=4.9.0.80